requirements.txt
python-telegram-bot==20.6
gspread
oauth2client
python-dotenv
openai
